move under styles/utils folder
