+++ src/styles/utils/generators/TooltipStyleUtils.ts
@@ -1,7 +1,7 @@
 import type {TextStyle, View, ViewStyle} from 'react-native';
 import {Animated} from 'react-native';
 import roundToNearestMultipleOfFour from '@libs/roundToNearestMultipleOfFour';
-import titleBarHeight from '@styles/titleBarHeight';
+import titleBarHeight from '@styles/utils/titleBarHeight';
 import FontUtils from '@styles/utils/FontUtils';
 // eslint-disable-next-line no-restricted-imports
 import positioning from '@styles/utils/positioning';
